Fix todos.
----

Cursor: “I’ve updated the todos alias to use a simpler pattern that matches any characters after the checkbox. Instead of [^\w\s] which was causing issues with Unicode characters, we now use .* to match any characters after the checkbox. This should properly display emojis and other Unicode characters in the output.”